+++ pydidas/widgets/workflow_edit/plugin_collection_browser.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ plugins to add them to the WorkflowTree.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -134,6 +134,6 @@ class PluginCollectionBrowser(CreateWidgetsMixIn, EmptyWidget):
         ]:
             return
         _p = self.collection.get_plugin_by_plugin_name(name)
-        self._widgets["plugin_description"].set_text_from_dict(
-            _p.get_class_description_as_dict(), _p.plugin_name
+        self._widgets["plugin_description"].set_text_from_list(
+            _p.get_class_description_as_list(), _p.plugin_name
         )