+++ CHANGELOG.rst
@@ -31,6 +31,8 @@ Improvements
     - Updated WorkflowResults contexts to accept dynamic result shapes.
     - Removed redundant code from plugins (because of changes in shape handling).
     - Switched to using ruff instead of black, flake8 and isort in github actions.
+- Improved the formatting when displaying Plugin information.
+
 
 Bugfixes
 --------